fix(TDQ-16642) Can preview unsigned int '4294967291' (#470)

+++ main/plugins/net.sourceforge.sqlexplorer/src/net/sourceforge/sqlexplorer/dataset/DataSet.java
@@ -419,7 +419,7 @@ public class DataSet {
         case Types.INTEGER:
         case Types.SMALLINT:
         case Types.TINYINT:
-            return new Long(resultSet.getInt(columnIndex));
+            return new Long(resultSet.getLong(columnIndex));
 
         case Types.BIGINT:
             return new Long(resultSet.getLong(columnIndex));
